Forbid deleting a cluster in creating state
If create and delete playbooks (post and delete API requests) are
launched immediately after each other, resources in OpenStack will be
allocated for a cluster that no longer exists in the API because delete
is much faster. That leaves allocated resources in OS that cannot be
viewed/deleted via RHub.

+++ src/rhub/api/lab/cluster.py
@@ -478,6 +478,13 @@ def delete_cluster(cluster_id, user):
         return problem(400, 'Bad Request',
                        f'Cluster {cluster_id} was already deleted')
 
+    if cluster.status.is_creating:
+        return problem(
+            400, 'Bad Request',
+            f'Cluster {cluster_id} is in creating state. Before deleting, '
+            'the cluster must be in the Active state or in any of failed states.',
+        )
+
     try:
         lab_utils.delete_cluster(cluster, user)
     except Exception:
